Cambios en el carrito y perfil, daba error

+++ src/app/carrito/carrito.ts
@@ -8,18 +8,28 @@ import { Cart, CartItem } from '../cart';
   styleUrl: './carrito.scss',
 })
 export class Carrito implements OnInit {
-  cartItems: CartItem [] = [];
+  cartItems: CartItem[] = [];
   errorMessage = '';
 
-  constructor(private cartService: Cart) {}
+  constructor(private cartService: Cart) { }
 
   ngOnInit(): void {
     this.cartItems = this.cartService.getCart();
   }
 
-  remove(item: CartItem){
+  remove(item: CartItem) {
+
+    if (this.cartItems.length === 0) {
+      this.errorMessage = 'El carrito ya está vacío';
+      return;
+    }
+
     this.cartService.removeFromCart(item.id, item.type);
+
     this.cartItems = this.cartService.getCart();
+
+    this.errorMessage = 'Libro eliminado del carrito';
+
   }
 
 }
